requirements.txt
streamlit
langchain
faiss-cpu
pymupdf
sentence-transformers
tqdm
ollama
langchain-community
langchain-ollama


python -m venv env

.\env\Scripts\activate

ollama pull gemma:2b
ollama run gemma:2b
streamlit run app.py
